+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Improve some SQL Fetcher log messages.
+
 ## [0.2.0] - 2022-11-30
 
 ### Fixed